develop:코딩 정리 및 보이는 정리 마무리

+++ articles/views.py
@@ -12,9 +12,6 @@ from rest_framework import generics
 from django.db.models import Q
 
 class ArticleListAPIView(APIView):
-
-    # permission_classes = [IsAuthenticated]
-
     def get(self, request):
         category = request.query_params.get('category')
         if category:
@@ -25,18 +22,15 @@ class ArticleListAPIView(APIView):
         return Response(serializer.data)
 
     def post(self, request):
-        serializer = ArticleSerializer(data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save(author=request.user)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)    
-        
-        
+        if request.user.is_authenticated:
+            serializer = ArticleSerializer(data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save(author=request.user)
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response({"error": "로그인이 필요합니다."}, status=400)
         
 class ArticleDetailAPIView(APIView):
-    
-    
-    # permission_classes = [IsAuthenticated]
-		
     def get_object(self, article_pk):
         return get_object_or_404(Article, pk=article_pk)
 
@@ -47,23 +41,30 @@ class ArticleDetailAPIView(APIView):
 
     def put(self, request, article_pk):
         article = self.get_object(article_pk)
-        serializer = ArticleDetailSerializer(article, data=request.data, partial=True)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.data)
-
+        if request.user.is_authenticated:
+            if request.user==article.author:
+                serializer = ArticleDetailSerializer(article, data=request.data, partial=True)
+                if serializer.is_valid(raise_exception=True):
+                    serializer.save()
+                    return Response(serializer.data, status=status.HTTP_201_CREATED)
+                return Response(serializer.data)
+            return Response({"message":"권한이 없습니다"})
+        else:
+            return Response({"error": "로그인이 필요합니다."}, status=400)
+    
     def delete(self, request, article_pk):
         article = self.get_object(article_pk)
-        article.delete()
-        data = {"pk": f"{article_pk} is deleted."}
-        return Response(data, status=status.HTTP_200_OK)
-    
+        if request.user.is_authenticated:
+            if request.user == article.author:
+                article.delete()
+                data = {"message": "게시글이 삭제되었습니다"}
+                return Response(data, status=status.HTTP_200_OK)
+            return Response({"message":"권한이 없습니다"})
+        else:
+            return Response({"error": "로그인이 필요합니다."}, status=400)
     
 class ArticleLikeAPIView(APIView):
-    
     permission_classes = [IsAuthenticated]
-    
     def put(self, request, article_pk):
         article = get_object_or_404(Article, pk=article_pk)
         article.likes.add(request.user)
@@ -72,11 +73,6 @@ class ArticleLikeAPIView(APIView):
         article = get_object_or_404(Article, pk=article_pk)
         article.likes.remove(request.user)
         return Response({"message": "Unliked"}, status=status.HTTP_200_OK)
-    
-    
-    
-    
-    
     
 class CommentView(APIView):
     # 댓글 조회 나중에 article 조회랑 합칠 것
@@ -88,19 +84,23 @@ class CommentView(APIView):
 
     # 댓글 작성
     def post(self, request, article_pk):
-        content = request.data.get('content')
-        article = get_object_or_404(Article, pk=article_pk)
-        comment = Comment.objects.create(
-            content=content,
-            author=request.user,
-            article=article,
-        )
-        serializer = CommentSerializer(comment, data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        if request.user.is_authenticated:
+            content = request.data.get('content')
+            article = get_object_or_404(Article, pk=article_pk)
+            comment = Comment.objects.create(
+                content=content,
+                author=request.user,
+                article=article,
+            )
+            serializer = CommentSerializer(comment, data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response({"error": "로그인이 필요합니다."}, status=400)
 
 class CommentDetailView(APIView):
+    permission_classes = [IsAuthenticated]
     # 댓글 수정
     def put(self, request, article_pk, comment_pk):
         comment = get_object_or_404(Comment, pk=comment_pk)
@@ -110,15 +110,15 @@ class CommentDetailView(APIView):
                 serializer.save()
                 return Response(serializer.data, status=status.HTTP_200_OK)
         else:
-            return Response({"error": "You do not have permission to edit this comment."}, status=400)
+            return Response({"error": "권한이 없습니다"}, status=400)
     # 댓글 삭제
     def delete(self, request, article_pk, comment_pk):
         comment = get_object_or_404(Comment, pk=comment_pk)
         if comment.author == request.user:
             comment.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
+            return Response({"mssege":"댓글이 삭제되었습니다"},status=status.HTTP_204_NO_CONTENT)
         else:
-            return Response({"error": "You do not have permission to delete this comment."}, status=400)
+            return Response({"error": "권한이 없습니다"}, status=400)
 
 class CommentFavoriteView(APIView):
     def post(self, request, article_pk, comment_pk):
@@ -131,7 +131,7 @@ class CommentFavoriteView(APIView):
             serializer = CommentSerializer(comment)
             return Response(serializer.data, status=200)
         else:
-            return Response({"error": "need to login."}, status=400)
+            return Response({"error": "로그인이 필요합니다."}, status=400)
 
 class SearchView(generics.ListAPIView):
     serializer_class = ArticleSerializer
@@ -141,7 +141,6 @@ class SearchView(generics.ListAPIView):
         title = query_params.get("title")
         content = query_params.get("content")
         author = query_params.get("author")
-
         q = Q()
         if title:
             q |= Q(title__icontains=title)
@@ -149,5 +148,4 @@ class SearchView(generics.ListAPIView):
             q |= Q(content__icontains=content)
         if author:
             q |= Q(author__username=author)
-
         return Article.objects.filter(q)
